Added a slew of flags from db.h
The flags I added are pretty arbitrary...if a block of #defines
was prefixed by "private" or something similar, or if a block
appeared inside a struct, I typically did not put the flags in.
If I'm missing a flag, or if I included a flag I shouldn't have,
please add/remove it and send me a patch.